+++ src/components/TextPressable/index.ts
@@ -1,5 +1,8 @@
+import React from 'react';
 import {TextProps} from 'react-native';
 
 export {default} from './TextPressable';
 
-export interface TextPressableProps extends TextProps {}
+export interface TextPressableProps extends TextProps {
+    children: React.ReactNode
+}