+++ src/containers/Profile/Header/Header.js
@@ -1,20 +1,22 @@
-import React, { Component } from 'react';
-import { StyleSheet, View, Text, TouchableOpacity } from 'react-native';
+import React, {Component} from 'react';
+import {StyleSheet, View, Text, TouchableOpacity} from 'react-native';
 import appConfig from 'app-config';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import AntIcon from 'react-native-vector-icons/AntDesign';
 import SVGLandscape from 'src/images/landscape.svg';
-import { CachedImageBackground } from 'react-native-img-cache';
+import {CachedImageBackground} from 'react-native-img-cache';
 import ImagePicker from 'react-native-image-picker';
 import ProfileContext from '../ProfileContext';
 import PremiumIcon from 'src/components/PremiumIcon';
+import {TouchableHighlight} from 'react-native-gesture-handler';
+import Loading from 'src/components/Loading';
 
 class Header extends Component {
   static contextType = ProfileContext;
   state = {};
 
   onChangeCover = () => {
-    const { t } = this.props;
+    const {t} = this.props;
 
     const options = {
       title: t('coverPicker.title'),
@@ -24,11 +26,11 @@ class Header extends Component {
       rotation: 360,
       storageOptions: {
         skipBackup: true,
-        path: 'images'
-      }
+        path: 'images',
+      },
     };
 
-    ImagePicker.showImagePicker(options, response => {
+    ImagePicker.showImagePicker(options, (response) => {
       if (response.error) {
         console.log(response.error);
       } else if (response.didCancel) {
@@ -44,7 +46,7 @@ class Header extends Component {
   };
 
   render() {
-    const { isMainUser } = this.context;
+    const {isMainUser} = this.context;
     return (
       <View style={styles.container}>
         {/* {isMainUser && (
@@ -52,36 +54,47 @@ class Header extends Component {
         )} */}
         <CachedImageBackground
           mutable
-          source={{ uri: this.props.cover }}
-          style={[styles.wrapper]}
-        >
+          source={{uri: this.props.cover}}
+          style={[styles.wrapper]}>
           <View style={styles.defaultCover}>
-            {!!!this.props.cover && !this.props.loading && (
+            {!!!this.props.cover && (
               <SVGLandscape
                 fill={appConfig.colors.primary}
                 height={'100%'}
                 width={'100%'}
-                style={{ position: 'absolute' }}
+                style={{position: 'absolute'}}
               />
             )}
 
             {isMainUser && (
               <TouchableOpacity
                 style={styles.cameraIconContainer}
-                onPress={this.onChangeCover}
-              >
+                onPress={this.onChangeCover}>
                 <AntIcon name="camerao" style={styles.cameraIcon} />
               </TouchableOpacity>
             )}
           </View>
           <View style={styles.avatarContainer}>
-            <View style={styles.avatarWrapper}>
-              <CachedImage
-                mutable
-                source={{ uri: this.props.avatar }}
-                style={styles.avatar}
-              />
-            </View>
+            <TouchableHighlight
+              disabled={this.props.avatarLoading}
+              underlayColor="rgba(0,0,0,.3)"
+              onPress={this.props.onPressAvatar}
+              style={styles.avatarWrapper}>
+              <View style={styles.avatar}>
+                <CachedImage
+                  mutable
+                  source={{uri: this.props.avatar}}
+                  style={styles.avatar}
+                />
+                {!!this.props.avatarLoading && (
+                  <Loading
+                    wrapperStyle={styles.avatarLoadingWrapper}
+                    center
+                    size="small"
+                  />
+                )}
+              </View>
+            </TouchableHighlight>
             {!!this.props.premium && (
               <PremiumIcon
                 premium={this.props.premium}
@@ -114,7 +127,7 @@ class Header extends Component {
 const styles = StyleSheet.create({
   container: {
     backgroundColor: '#fafafa',
-    paddingBottom: 15
+    paddingBottom: 15,
   },
   wrapper: {
     width: appConfig.device.width,
@@ -122,27 +135,31 @@ const styles = StyleSheet.create({
     backgroundColor: '#b5b5b5',
     alignItems: 'center',
     justifyContent: 'flex-end',
-    marginBottom: 20,
+    marginBottom: 33,
     borderBottomWidth: 3,
-    borderBottomColor: '#eee'
+    borderBottomColor: '#eee',
   },
   avatarContainer: {
-    width: 92,
-    height: 92,
-    bottom: -20
+    width: 100,
+    height: 100,
+    bottom: -33,
+    backgroundColor: '#eee',
+    borderRadius: 50,
   },
   avatarWrapper: {
-    backgroundColor: '#eee',
     alignItems: 'center',
-    borderRadius: 46,
+    borderRadius: 50,
     overflow: 'hidden',
     borderWidth: 3,
-    borderColor: '#eee'
+    borderColor: '#eee',
   },
   avatar: {
     width: '100%',
     height: '100%',
-    resizeMode: 'cover'
+    resizeMode: 'cover',
+  },
+  avatarLoadingWrapper: {
+    backgroundColor: 'rgba(0,0,0,.6)',
   },
   premium: {
     position: 'absolute',
@@ -150,7 +167,7 @@ const styles = StyleSheet.create({
     right: 10,
     backgroundColor: '#fafafa',
     padding: 4,
-    borderRadius: 15
+    borderRadius: 15,
   },
   title: {
     textAlign: 'center',
@@ -159,7 +176,7 @@ const styles = StyleSheet.create({
     fontSize: 18,
     marginTop: 10,
     letterSpacing: 1,
-    paddingHorizontal: 15
+    paddingHorizontal: 15,
   },
   subTitle: {
     textAlign: 'center',
@@ -167,64 +184,64 @@ const styles = StyleSheet.create({
     marginTop: 3,
     fontWeight: '300',
     fontStyle: 'italic',
-    paddingHorizontal: 15
+    paddingHorizontal: 15,
   },
   infoBlock: {
     flexDirection: 'row',
     marginTop: 10,
     borderTopColor: '#eee',
-    borderTopWidth: 0.5
+    borderTopWidth: 0.5,
   },
   row: {
     flexDirection: 'row',
     alignItems: 'center',
     paddingHorizontal: 15,
-    marginTop: 15
+    marginTop: 15,
   },
   icon: {
     width: 20,
     fontSize: 16,
     marginRight: 0,
-    color: '#444'
+    color: '#444',
   },
   rowTitle: {
     color: '#444',
     fontSize: 13,
     lineHeight: 18,
-    flex: 1
+    flex: 1,
   },
   iconDistance: {
     color: '#888',
     fontSize: 14,
-    flex: 0
+    flex: 0,
   },
   distance: {
     position: 'absolute',
     bottom: 15,
-    right: 0
+    right: 0,
     // alignItems: 'flex-end',
     // flex: 1
   },
   defaultCover: {
     position: 'absolute',
     width: '100%',
-    height: '100%'
+    height: '100%',
   },
   cameraIconContainer: {
     ...elevationShadowStyle(7),
     position: 'absolute',
     right: 10,
-    top: appConfig.isIOS ? 60 : 75
+    top: appConfig.isIOS ? 60 : 75,
   },
   cameraIcon: {
     fontSize: 26,
-    color: '#fff'
-  }
+    color: '#fff',
+  },
 });
 
 export default withTranslation('editProfile')(Header);
 
-const Row = props => (
+const Row = (props) => (
   <View style={[styles.row, props.style]}>
     <Icon name={props.iconName} style={[styles.icon, props.iconStyle]} />
     <Text style={[styles.rowTitle, props.titleStyle]}>{props.content}</Text>