+++ src/containers/Profile/Profile.js
@@ -1,12 +1,7 @@
-import React, { Component } from 'react';
-import {
-  StyleSheet,
-  Animated,
-  Alert,
-  View,
-  RefreshControl
-} from 'react-native';
-import { Actions } from 'react-native-router-flux';
+import React, {Component} from 'react';
+import {StyleSheet, Animated, Alert, View, RefreshControl} from 'react-native';
+import {Actions} from 'react-native-router-flux';
+import ImagePicker from 'react-native-image-picker';
 import store from 'app-store';
 import appConfig from 'app-config';
 import Header from './Header';
@@ -14,7 +9,7 @@ import ComboButton from './ComboButton';
 import Gallery from './Gallery';
 import Loading from 'src/components/Loading';
 import RNFetchBlob from 'rn-fetch-blob';
-import { ProfileProvider } from './ProfileContext';
+import {ProfileProvider} from './ProfileContext';
 import Communications from 'react-native-communications';
 import Intro from './Intro';
 import NavBar from './NavBar';
@@ -22,50 +17,51 @@ import EventTracker from 'src/helper/EventTracker';
 
 const IMAGE_TYPE = {
   COVER: 'cover',
-  IMAGE: 'image'
+  IMAGE: 'image',
 };
 
 class Profile extends Component {
   static defaultProps = {
-    isMainUser: false
+    isMainUser: false,
   };
   state = {
     userInfo: {},
     gallery: null,
     loading: true,
     refreshing: false,
+    avatarLoading: false,
     upperLayout: undefined,
-    animatedScroll: new Animated.Value(0)
+    animatedScroll: new Animated.Value(0),
   };
   unmounted = false;
   eventTracker = new EventTracker();
 
   get introData() {
-    const { userInfo } = this.state;
+    const {userInfo} = this.state;
     let data = [];
     if (userInfo.premium != MEMBERSHIP_TYPE.STANDARD) {
       data = data.concat([
         {
           title: 'Ngành nghề',
-          value: userInfo.name_profession
+          value: userInfo.name_profession,
         },
         {
           title: 'Thương hiệu',
-          value: userInfo.brand
-        }
+          value: userInfo.brand,
+        },
       ]);
     }
     data = data.concat([
       {
         title: 'My Facebook',
         select: true,
-        value: userInfo.facebook
+        value: userInfo.facebook,
       },
       {
         title: 'My Youtube',
         select: true,
-        value: userInfo.youtube
-      }
+        value: userInfo.youtube,
+      },
     ]);
 
     return data;
@@ -89,22 +85,23 @@ class Profile extends Component {
 
     const user_id = userInfo.id;
 
-    const { store_id } = store;
-    let { t } = this.props;
+    const {store_id} = store;
+    let {t} = this.props;
     if (store_id === undefined || store_id === null) return;
 
     try {
       const response = await APIHandler.site_user_profile(store_id, user_id);
+      console.log(response)
       if (!this.unmounted) {
         if (response && response.status === STATUS_SUCCESS && response.data) {
           this.setState({
             userInfo: response.data.user,
-            gallery: response.data.images || []
+            gallery: response.data.images || [],
           });
         } else {
           flashShowMessage({
             type: 'danger',
-            message: response.message || t('common:api.error.message')
+            message: response.message || t('common:api.error.message'),
           });
         }
       }
@@ -112,22 +109,22 @@ class Profile extends Component {
       console.log('site_user_profile', error);
       flashShowMessage({
         type: 'danger',
-        message: t('common:api.error.message')
+        message: t('common:api.error.message'),
       });
     } finally {
-      !this.unmounted && this.setState({ loading: false, refreshing: false });
+      !this.unmounted && this.setState({loading: false, refreshing: false});
     }
   };
 
   handleEdit = () => {
     Actions.push(appConfig.routes.editPersonalProfile, {
       user_info: this.state.userInfo,
-      refresh: this.getProfile
+      refresh: this.getProfile,
     });
   };
 
   handleLogout = () => {
-    const { t } = this.props;
+    const {t} = this.props;
 
     Alert.alert(
       t('signOut.warningTitle'),
@@ -135,28 +132,27 @@ class Profile extends Component {
       [
         {
           text: t('signOut.cancel'),
-          onPress: () => {}
+          onPress: () => {},
         },
         {
           text: t('signOut.title'),
           onPress: this.logout,
-          style: 'destructive'
-        }
+          style: 'destructive',
+        },
       ],
-      { cancelable: false }
+      {cancelable: false},
     );
   };
 
   logout = async () => {
     this.setState({
-      loading: true
+      loading: true,
     });
     try {
-      const { t } = this.props;
+      const {t} = this.props;
       const response = await APIHandler.user_logout();
       switch (response.status) {
         case STATUS_SUCCESS:
-          EventTracker.removeUserId();
           store.setUserInfo(response.data);
           store.resetCartData();
           store.setRefreshHomeChange(store.refresh_home_change + 1);
@@ -164,9 +160,9 @@ class Profile extends Component {
           store.resetAsyncStorage();
           flashShowMessage({
             message: t('account:signOut.successMessage'),
-            type: 'info'
+            type: 'info',
           });
-          Actions.reset(appConfig.routes.sceneWrapper);
+          setTimeout(() => Actions.reset(appConfig.routes.sceneWrapper));
           break;
         default:
           console.log('default');
@@ -175,7 +171,7 @@ class Profile extends Component {
       console.log(error);
     } finally {
       this.setState({
-        loading: false
+        loading: false,
       });
     }
   };
@@ -195,21 +191,21 @@ class Profile extends Component {
       Alert.alert(
         'Không thể liên lạc',
         userName + ' chưa đăng ký số điện thoại',
-        [{ text: 'Đã hiểu' }]
+        [{text: 'Đã hiểu'}],
       );
     }
   };
 
   onChat = async () => {
     this.setState({
-      loading: true
+      loading: true,
     });
-    const { t } = this.props;
+    const {t} = this.props;
     if (!this.state.userInfo) return;
     if (!this.state.userInfo.id) return;
 
     const data = {
-      friend_id: this.state.userInfo.id
+      friend_id: this.state.userInfo.id,
     };
     try {
       const response = await APIHandler.user_create_conversation(data);
@@ -218,30 +214,30 @@ class Profile extends Component {
         Actions.push(appConfig.routes.amazingUserChat, {
           user: response.data.user,
           title: response.data.user.name,
-          conversation_id: response.data.conversation_id
+          conversation_id: response.data.conversation_id,
         });
       } else {
         flashShowMessage({
           type: 'danger',
-          message: response.message || t('api.error.message')
+          message: response.message || t('api.error.message'),
         });
       }
     } catch (error) {
       console.log('create_conversation', error);
       flashShowMessage({
         type: 'danger',
-        message: t('api.error.message')
+        message: t('api.error.message'),
       });
     } finally {
       !this.unmounted &&
         this.setState({
-          loading: false
+          loading: false,
         });
     }
   };
 
-  uploadMultiTempImage = async images => {
-    this.setState({ loading: true });
+  uploadMultiTempImage = async (images) => {
+    this.setState({loading: true});
     const uploadImages = [];
     for (const image of images) {
       const img = await this.uploadTempImageLoop(image);
@@ -255,8 +251,8 @@ class Profile extends Component {
       this.uploadTempImage(
         image,
         IMAGE_TYPE.IMAGE,
-        data => resolve(data),
-        () => reject()
+        (data) => resolve(data),
+        () => reject(),
       );
     });
   }
@@ -278,27 +274,27 @@ class Profile extends Component {
   uploadTempImage = (response, type, resolve, reject) => {
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       () => {
-        const { t } = this.props;
+        const {t} = this.props;
         const image = {
           name: 'image',
           filename: response.filename,
-          data: response.data
+          data: response.data,
         };
         // call api post my form data
         RNFetchBlob.fetch(
           'POST',
           APIHandler.url_user_upload_image(),
           {
-            'Content-Type': 'multipart/form-data'
+            'Content-Type': 'multipart/form-data',
           },
-          [image]
+          [image],
         )
-          .then(resp => {
+          .then((resp) => {
             if (!this.unmounted) {
-              const { data } = resp;
+              const {data} = resp;
               const response = JSON.parse(data);
               if (
                 response &&
@@ -314,40 +310,40 @@ class Profile extends Component {
                 reject && reject();
                 flashShowMessage({
                   type: 'danger',
-                  message: response.message || t('common:api.error.message')
+                  message: response.message || t('common:api.error.message'),
                 });
-                !this.unmounted && this.setState({ loading: false });
+                !this.unmounted && this.setState({loading: false});
               }
             }
           })
-          .catch(error => {
+          .catch((error) => {
             reject && reject();
 
             console.log(error);
             flashShowMessage({
               type: 'danger',
-              message: t('common:api.error.message')
+              message: t('common:api.error.message'),
             });
-            !this.unmounted && this.setState({ loading: false });
+            !this.unmounted && this.setState({loading: false});
           });
-      }
+      },
     );
   };
 
   async uploadImage(images) {
     const data = {
-      image_name: images
+      image_name: images,
     };
-    const { t } = this.props;
+    const {t} = this.props;
     try {
       const response = await APIHandler.user_upload_image(data);
       if (!this.unmounted) {
         if (response && response.status === STATUS_SUCCESS && response.data) {
-          this.setState({ gallery: response.data.images });
+          this.setState({gallery: response.data.images});
         } else {
           flashShowMessage({
             type: 'danger',
-            message: response.message || t('common:api.error.message')
+            message: response.message || t('common:api.error.message'),
           });
         }
       }
@@ -355,40 +351,40 @@ class Profile extends Component {
       console.log('user_upload_image', error);
       flashShowMessage({
         type: 'danger',
-        message: t('common:api.error.message')
+        message: t('common:api.error.message'),
       });
     } finally {
       !this.unmounted &&
         this.setState({
-          loading: false
+          loading: false,
         });
     }
   }
 
-  deleteImage = async images => {
-    this.setState({ loading: true });
+  deleteImage = async (images) => {
+    this.setState({loading: true});
     if (!Array.isArray(images)) {
       images = [images];
     }
     const data = {
-      user_image_ids: images
+      user_image_ids: images,
     };
 
-    const { t } = this.props;
+    const {t} = this.props;
     try {
       const response = await APIHandler.user_delete_image(data);
 
       if (!this.unmounted) {
         if (response && response.status === STATUS_SUCCESS && response.data) {
-          this.setState({ gallery: response.data.images || [] });
+          this.setState({gallery: response.data.images || []});
           flashShowMessage({
             type: 'success',
-            message: response.message
+            message: response.message,
           });
         } else {
           flashShowMessage({
             type: 'danger',
-            message: response.message || t('common:api.error.message')
+            message: response.message || t('common:api.error.message'),
           });
         }
       }
@@ -396,30 +392,119 @@ class Profile extends Component {
       console.log('user_delete_image', error);
       flashShowMessage({
         type: 'danger',
-        message: t('common:api.error.message')
+        message: t('common:api.error.message'),
       });
     } finally {
       !this.unmounted &&
         this.setState({
-          loading: false
+          loading: false,
         });
     }
   };
 
+  onTapAvatar = () => {
+    const {t} = this.props;
+
+    const options = {
+      cameraType: 'front',
+      rotation: 360,
+      title: t('account:avatarPicker.title'),
+      cancelButtonTitle: t('account:avatarPicker.cancelTitle'),
+      takePhotoButtonTitle: t('account:avatarPicker.takePhotoTitle'),
+      chooseFromLibraryButtonTitle: t(
+        'account:avatarPicker.chooseFromLibraryTitle',
+      ),
+      storageOptions: {
+        // skipBackup: true,
+        path: 'images',
+      },
+    };
+
+    ImagePicker.showImagePicker(options, (response) => {
+      if (response.error) {
+        console.log(response.error);
+      } else if (response.didCancel) {
+        console.log(response);
+      } else {
+        if (!response.fileName) {
+          response.fileName = new Date().getTime();
+          if (response.type) {
+            response.fileName += '.' + response.type.split('image/')[1];
+          } else {
+            response.fileName += '.jpeg';
+          }
+        }
+        this.uploadAvatar(response);
+      }
+    });
+  }
+
+  uploadAvatar(response) {
+    this.setState(
+      {
+        avatarLoading: true,
+      },
+      () => {
+        const avatar = {
+          name: 'avatar',
+          filename: response.fileName,
+          data: response.data,
+        };
+        // call api post my form data
+        RNFetchBlob.fetch(
+          'POST',
+          APIHandler.url_user_add_avatar(),
+          {
+            'Content-Type': 'multipart/form-data',
+          },
+          [avatar, {name: 'site_id', data: store.store_data?.id}],
+        )
+          .then((resp) => {
+            if (this.unmounted) return;
+
+            var {data} = resp;
+            var response = JSON.parse(data);
+            if (response) {
+              flashShowMessage({
+                type: response?.status == STATUS_SUCCESS ? 'success' : 'danger',
+                message:
+                  response?.message ||
+                  (response?.status === STATUS_SUCCESS &&
+                    this.props.t('common:api.error.message')),
+              });
+            }
+          })
+          .catch((error) => {
+            console.log(error);
+            flashShowMessage({
+              type: 'danger',
+              message: this.props.t('common:api.error.message'),
+            });
+          })
+          .finally(() => {
+            if (this.unmounted) return;
+            this.setState({
+              avatarLoading: false,
+            });
+          });
+      },
+    );
+  }
+
   async uploadCover(cover) {
     const data = {
-      image_cover: cover
+      image_cover: cover,
     };
-    const { t } = this.props;
+    const {t} = this.props;
     try {
       const response = await APIHandler.user_upload_image_cover(data);
       if (!this.unmounted) {
         if (response && response.status === STATUS_SUCCESS && response.data) {
-          this.setState({ userInfo: response.data.user });
+          this.setState({userInfo: response.data.user});
         } else {
           flashShowMessage({
             type: 'danger',
-            message: response.message || t('common:api.error.message')
+            message: response.message || t('common:api.error.message'),
           });
         }
       }
@@ -427,22 +512,22 @@ class Profile extends Component {
       console.log('user_upload_image', error);
       flashShowMessage({
         type: 'danger',
-        message: t('common:api.error.message')
+        message: t('common:api.error.message'),
       });
     } finally {
       !this.unmounted &&
         this.setState({
-          loading: false
+          loading: false,
         });
     }
   }
 
-  onUpperLayout = e => {
-    this.setState({ upperLayout: e.nativeEvent.layout.height });
+  onUpperLayout = (e) => {
+    this.setState({upperLayout: e.nativeEvent.layout.height});
   };
 
   onRefresh = () => {
-    this.setState({ refreshing: true });
+    this.setState({refreshing: true});
     this.getProfile();
   };
 
@@ -451,12 +536,13 @@ class Profile extends Component {
       isMainUser: this.props.isMainUser,
       upperLayout: this.state.upperLayout,
       animatedScroll: this.state.animatedScroll,
-      animatedInputRange: [0, 70]
+      animatedInputRange: [0, 70],
     };
 
     return (
       <ProfileProvider value={data}>
         <NavBar
+          onChat={this.onChat}
           onEdit={this.handleEdit}
           onLogout={this.handleLogout}
           title={this.state.userInfo.name}
@@ -467,12 +553,12 @@ class Profile extends Component {
               {
                 nativeEvent: {
                   contentOffset: {
-                    y: this.state.animatedScroll
-                  }
-                }
-              }
+                    y: this.state.animatedScroll,
+                  },
+                },
+              },
             ],
-            { useNativeDriver: true }
+            {useNativeDriver: true},
           )}
           style={styles.container}
           scrollEventThrottle={16}
@@ -481,25 +567,26 @@ class Profile extends Component {
               refreshing={this.state.refreshing}
               onRefresh={this.onRefresh}
             />
-          }
-        >
+          }>
           {this.state.loading && <Loading center />}
           <View onLayout={this.onUpperLayout}>
             <Header
-              onEdit={this.handleEdit}
-              onLogout={this.handleLogout}
+              avatarLoading={this.state.avatarLoading}
               avatar={this.state.userInfo.avatar}
               cover={this.state.userInfo.image_cover}
               name={this.state.userInfo.name}
               quote={this.state.userInfo.quote}
               address={this.state.userInfo.address_view}
-              tel={this.state.userInfo.tel}
+              // tel={this.state.userInfo.tel}
               distance={this.state.userInfo.distance}
               premium={this.state.userInfo.premium}
               loading={this.state.loading}
-              uploadTempCover={data =>
+              uploadTempCover={(data) =>
                 this.uploadTempImage(data, IMAGE_TYPE.COVER)
               }
+              onPressAvatar={this.onTapAvatar}
+              onEdit={this.handleEdit}
+              onLogout={this.handleLogout}
             />
           </View>
           <Gallery
@@ -507,15 +594,15 @@ class Profile extends Component {
             uploadTempImage={this.uploadTempImage}
             onDeleteImage={this.deleteImage}
             data={this.state.gallery}
-            headerComponent={
-              this.props.isMainUser || (
-                <ComboButton
-                  style={styles.comboBtn}
-                  onCall={this.onCall}
-                  onChat={this.onChat}
-                />
-              )
-            }
+            // headerComponent={
+            //   this.props.isMainUser || (
+            //     <ComboButton
+            //       style={styles.comboBtn}
+            //       onCall={this.onCall}
+            //       onChat={this.onChat}
+            //     />
+            //   )
+            // }
           />
           <Intro content={this.state.userInfo.intro} data={this.introData} />
         </Animated.ScrollView>
@@ -531,17 +618,19 @@ const styles = StyleSheet.create({
     height: '100%',
     width: '100%',
     paddingBottom: appConfig.device.bottomSpace,
-    backgroundColor: '#fcfcfc'
+    backgroundColor: '#fcfcfc',
   },
   icon: {
     color: '#fff',
     fontSize: 22,
     paddingHorizontal: 10,
-    ...elevationShadowStyle(7)
+    ...elevationShadowStyle(7),
   },
   comboBtn: {
-    backgroundColor: '#f9f9f9'
-  }
+    backgroundColor: '#f9f9f9',
+  },
 });
 
-export default withTranslation(['profileDetail', 'common'])(observer(Profile));
+export default withTranslation(['profileDetail', 'common', 'account'])(
+  observer(Profile),
+);