+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [r12.1.1] - 2021-08-16
+## [r12.1.1] - 2021-08-17
 ### Added
 - [CU-cquwh3] Function abaha_request and request_management in serviceHandler by [lamah.dev@gmail.com].
 - [CU-bv7k11] Scan QR code button when enter invitation code by [kienluu612@gmail.com] [minhnguyenit14@gmail.com].
@@ -15,11 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CU-aty6rq] LoadMore to show missing products in GroupProduct by [nguyenanhtuan191298@gmail.com].
 - [CU-9227a2][CU-bng0v8] Notifications in tab bar by [minhnguyenit14@gmail.com].
 - [CU-anxf53] Copy button when press long array numbers in GiftedChat view  by [kienluu612@gmail.com] [minhnguyenit14@gmail.com].
+- voucher_campaign_detail in servicesHandler by [minhnguyenit14@gmail.com].
+- [CU-cqx6x5] Avatar Image of ItemAttribute is not changed while choosing only 1 attribute by [kienluu612@gmail.com] [minhnguyenit14@gmail.com].
 ### Changed
 - [CU-akxhu3] Zoom image in chat screen by [lamah.dev@gmail.com].
 - [CU-ar0bbd] Prevent buying when product is out of stock by [minhnguyenit14@gmail.com].
 - [CU-akz4up] Using product's attrs, models instead of calling to api to get them by [minhnguyenit14@gmail.com].
 - [Android] compileSdkVersion, targetSdkVersion, buildToolsVersion by [minhnguyenit14@gmail.com].
+- Use Profile for Account detail by [minhnguyenit14@gmail.com].
 ### Fixed
 - [CU-bnknw5] View not-ordering orders when there is no address in cart_data by [lamah.dev@gmail.com].
 - Low performance because of using ScrollView in CategoryScreen by [minhnguyenit14@gmail.com].